Ability to create Wallet

+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/BroadcastedTransactionDesign.cs
@@ -9,7 +9,7 @@ public class BroadcastedTransactionDesign : IBroadcastedTransaction
     public string Address { get; init; }
     public decimal FeeRate { get; set; }
     public decimal TotalFee { get; set; }
-    public decimal Amount { get; init; }
+    public uint Amount { get; init; }
     public string Path { get; init; }
     public int UtxoCount { get; init; }
     public string ViewRawJson { get; init; }
